+++ pytorch_lightning/callbacks/lambda_function.py
