+++ tests/callbacks/test_lambda_function.py
